Move combat action utils
